reorganize code into a single file
